charts: split into modules
