+++ frontend/test/charts.test.ts
@@ -1,12 +1,11 @@
 import test from "ava";
 
+import { balances, commodities } from "../src/charts/line";
 import {
-  balances,
-  commodities,
   parseGroupedQueryChart,
   parseQueryChart,
-  scatterplot,
-} from "../src/charts";
+} from "../src/charts/query-charts";
+import { scatterplot } from "../src/charts/scatterplot";
 import type { Result } from "../src/lib/result";
 
 function force_ok<T>(r: Result<T, unknown>): T {
@@ -15,6 +14,8 @@ function force_ok<T>(r: Result<T, unknown>): T {
   }
   return r.value;
 }
+
+const ctx = { currencies: ["EUR"], dateFormat: () => "DATE" };
 
 test("handle data for balances chart", (t) => {
   t.false(balances("").success);
@@ -27,13 +28,13 @@ test("handle data for balances chart", (t) => {
   t.is(parsed.data.length, 2);
   t.is(parsed.data[0].values.length, 2);
   t.is(parsed.data[1].values.length, 1);
-  const queryChart = parseQueryChart(data, ["EUR"]);
+  const queryChart = parseQueryChart(data, ctx);
   t.deepEqual(queryChart.success && queryChart.value.data, parsed.data);
   t.snapshot(parsed);
 });
 
 test("handle data for commodities chart", (t) => {
-  t.false(commodities("", "asdf").success);
+  t.false(commodities("", null, "asdf").success);
   const data: unknown = {
     base: "EUR",
     quote: "USD",
@@ -42,7 +43,7 @@ test("handle data for commodities chart", (t) => {
       ["2000-02-01", 12.0],
     ],
   };
-  const parsed = force_ok(commodities(data, "asdf"));
+  const parsed = force_ok(commodities(data, null, "asdf"));
   t.is(parsed.data.length, 1);
   t.is(parsed.data[0].values.length, 2);
   t.snapshot(parsed);
@@ -60,7 +61,7 @@ test("handle data for scatterplot chart", (t) => {
 
 test("handle data for query charts", (t) => {
   const d = [{ group: "Assets:Cash", balance: { EUR: 10 } }];
-  const { data } = force_ok(parseGroupedQueryChart(d, ["EUR"]));
+  const { data } = force_ok(parseGroupedQueryChart(d, ctx));
   t.is(data.get("EUR")?.value, 10);
   t.deepEqual(
     data
@@ -73,6 +74,6 @@ test("handle data for query charts", (t) => {
 
 test("handle invalid data for query charts", (t) => {
   const d: unknown[] = [{}];
-  const c = parseQueryChart(d, []);
+  const c = parseQueryChart(d, ctx);
   t.false(c.success);
 });
